Make `DeserializationError` crate-private
